src/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Database
mysqlclient==2.2.1

# Background scheduler
apscheduler==3.10.4

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
pydantic==2.5.3

# AI/ML - Gemini API for Persona Synth
google-generativeai>=0.8.0
